requirements.txt
Flask==2.3.2
opencv-python
tensorflow
Pillow
Keras
scikit-learn
blinker==1.6.2  
click==8.1.4  
colorama==0.4.6  
importlib-metadata==6.8.0  
itsdangerous==2.1.2  
Jinja2==3.1.2  
MarkupSafe==2.1.3  
numpy==1.22.4 
pip==19.2.3 
setuptools==41.2.0 
Werkzeug==2.3.6  
zipp==3.16.0 